requirements.txt
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.8
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.1
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.16.0
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.1
zipp==3.1.0
